Fixed naming schema
There was an odd bug that came up in the naming schema where the ordering of the values would get messed up. I can go further into detail upon further questions. The way I was able to fix this was by starting every file name with a version number that starts with an a. This version will be 'a.01.somename'. After the database is completely filled, we can go back to a simpler 'version-less' naming schema, but for now (with constant update) we need the version schema.